Move validation function to projections file
